Worked on spells

+++ Core/Source/Base/SpellManager.cs
@@ -11,22 +11,22 @@ namespace OldWorldGods.Base
 {
     public class SpellManager
     {
-        public static void CastSpell(SpellDef spell, GodDef god, List<Pawn> casters, List<Thing> targets)
+        public static void CastSpell(SpellEffectDef effectDef, Thing rune, GodDef god, List<Pawn> casters, List<LocalTargetInfo> targets)
         {
-            foreach (SpellEffectDef effectDef in spell.effects)
+            if (targets.Count == 0)
             {
-                int totalTargets = (int) (effectDef.targets.baseValue * (1+effectDef.targets.casterMultiplier * casters.Count));
-                if(targets.Count > totalTargets) targets.RemoveRange(totalTargets, targets.Count);
-                float strength = CalculateNumber(effectDef.strength, casters.Count, targets.Count);
-                ApplyRecoil(casters, god, strength);
-                effectDef.Execute(god, strength, casters.Count, targets);
+                Messages.Message("SpellNoTargets".Translate(), MessageTypeDefOf.RejectInput);
+                return;
             }
+            int totalTargets =
+                (int)(effectDef.targets.baseValue * (1 + effectDef.targets.casterMultiplier * casters.Count));
+            if (targets.Count > totalTargets) targets.RemoveRange(totalTargets, targets.Count);
+            float strength = effectDef.strength.CalculateNumber(casters.Count, targets.Count);
+            ApplyRecoil(casters, god, strength);
+            effectDef.Execute(rune, god, strength, casters.Count, targets);
         }
 
-        private static float CalculateNumber(SpellNumber number, int casters, int targets)
-        {
-            return number.baseValue * (1 + number.casterMultiplier * casters) * (1 + number.targetMultiplier * targets);
-        }
+
         private static void ApplyRecoil(List<Pawn> casters, GodDef god, float strength)
         {
             foreach (Pawn caster in casters)
@@ -35,7 +35,8 @@ namespace OldWorldGods.Base
             }
         }
 
-        public static void ApplyGodEffect(GodDef god, Pawn pawn, int targets, int casters, float strength, DamageSpellDef damageSpell = null)
+        public static void ApplyGodEffect(GodDef god, Pawn pawn, int targets, int casters, float strength,
+            DamageSpellDef damageSpell = null)
         {
             switch (god.godEffect)
             {
@@ -44,8 +45,8 @@ namespace OldWorldGods.Base
                     if (damageSpell != null)
                     {
                         Hediff hediff = HediffMaker.MakeHediff(HediffDefOf.Cut, pawn);
-                        hediff.CurStage.deathMtbDays = 
-                            CalculateNumber(damageSpell.hoursLeft, casters, targets);
+                        hediff.CurStage.deathMtbDays =
+                            damageSpell.hoursLeft.CalculateNumber(casters, targets);
                         if (!damageSpell.kill) hediff.Severity = 0;
                         pawn.health.AddHediff(hediff, pawn.RaceProps.body.corePart,
                             new DamageInfo(DamageDefOf.Cut, strength, 9999,
@@ -57,16 +58,16 @@ namespace OldWorldGods.Base
                             new DamageInfo(DamageDefOf.Cut, strength, 9999,
                                 hitPart: pawn.RaceProps.body.corePart));
                     }
+
                     break;
                 }
                 case "poison":
                 {
-                    
                     if (damageSpell != null)
                     {
                         Hediff hediff = HediffMaker.MakeHediff(HediffDefOf.ToxicBuildup, pawn);
-                        hediff.CurStage.deathMtbDays = 
-                            CalculateNumber(damageSpell.hoursLeft, casters, targets);
+                        hediff.CurStage.deathMtbDays =
+                            damageSpell.hoursLeft.CalculateNumber(casters, targets);
                         if (!damageSpell.kill) hediff.Severity = 0;
                         pawn.health.AddHediff(hediff, pawn.RaceProps.body.corePart,
                             new DamageInfo(DamageDefOf.Rotting, strength, 9999,
@@ -78,13 +79,14 @@ namespace OldWorldGods.Base
                             new DamageInfo(DamageDefOf.Rotting, strength, 9999,
                                 hitPart: pawn.RaceProps.body.corePart));
                     }
+
                     break;
                 }
                 case "corruption":
                 {
                     if (pawn.Faction.Equals(Faction.OfPlayer))
                     {
-                        foreach (Pawn other in pawn.Map.mapPawns.FreeColonists.Where(other => other != pawn && 
+                        foreach (Pawn other in pawn.Map.mapPawns.FreeColonists.Where(other => other != pawn &&
                             pawn.relations.OpinionOf(other) > 0 && strength-- > 0))
                         {
                             other.needs.mood.thoughts.memories.TryGainMemory(RitualThoughtDefOf.RitualThought, pawn);
@@ -93,26 +95,29 @@ namespace OldWorldGods.Base
 
                     List<Thought> thoughts = new List<Thought>();
                     pawn.needs.mood.thoughts.GetAllMoodThoughts(thoughts);
-                    RitualThought target = (RitualThought) thoughts.Find(thought => thought is RitualThought);
+                    RitualThought target = (RitualThought)thoughts.Find(thought => thought is RitualThought);
                     if (target == null)
                     {
                         target = new RitualThought();
-                        pawn.needs.mood.thoughts.situational.AppendMoodThoughts(new List<Thought> { target });
+                        pawn.needs.mood.thoughts.situational.AppendMoodThoughts(new List<Thought> {target});
                     }
+
                     while (strength < 1)
                     {
                         target.SetStageIndex++;
                         strength /= 5;
                     }
+
                     if (target.CurStageIndex > target.def.stages.Count)
                     {
                         pawn.SetFaction(Find.FactionManager.FirstFactionOfDef(CorruptFactionDefOf.Corrupt));
                     }
+
                     break;
                 }
                 case "time":
                 {
-                    pawn.ageTracker.AgeTickMothballed((int) strength*300000);
+                    pawn.ageTracker.AgeTickMothballed((int)strength * 300000);
                     break;
                 }
             }
@@ -127,7 +132,7 @@ namespace OldWorldGods.Base
 
             if (itemEffect.itemHealthChange)
             {
-                apparel.HitPoints -= (int) strength;
+                apparel.HitPoints -= (int)strength;
             }
         }
     }
